requirements.txt
discord.py
pandas
scikit-learn
schedule
spacy
transformers
python-dotenv
openai